% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTranscriptAnnotation.R
\name{buildTranscriptAnnotation}
\alias{buildTranscriptAnnotation}
\title{Build transcript annotation from GTF}
\usage{
buildTranscriptAnnotation(gtf)
}
\arguments{
\item{gtf}{Path to a local GTF/GTF.GZ file.}
}
\value{
A list with components \code{exon} and \code{utr}.
}
\description{
Build transcript annotation from GTF
}
