% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFPKM.R
\name{computeFPKM}
\alias{computeFPKM}
\title{Compute FPKM from count matrix}
\usage{
computeFPKM(counts, geneLength)
}
\arguments{
\item{counts}{Gene by sample raw count matrix.}

\item{geneLength}{Named numeric vector of gene lengths in bp.}
}
\value{
A numeric FPKM matrix.
}
\description{
Compute FPKM from count matrix
}
