% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportResults.R
\name{exportResults}
\alias{exportResults}
\title{Export module results}
\usage{
exportResults(x, file, which = "result")
}
\arguments{
\item{x}{A result object.}

\item{file}{Output file path.}

\item{which}{Component to export.}
}
\description{
Export module results
}
