% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getASEGenes.R
\name{getASEGenes}
\alias{getASEGenes}
\title{Extract genes showing ASE}
\usage{
getASEGenes(x, sampleID = NULL)
}
\arguments{
\item{x}{An \code{ASEGeneResult} object.}

\item{sampleID}{Optional sample ID.}
}
\value{
A data.frame.
}
\description{
Extract genes showing ASE
}
