% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneAE.R
\name{getGeneAE}
\alias{getGeneAE}
\title{Extract AE results for one gene}
\usage{
getGeneAE(x, geneID, use = c("gene_id", "gene_name"))
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{geneID}{Gene identifier or gene symbol.}

\item{use}{One of \code{"gene_id"} or \code{"gene_name"}.}
}
\value{
A data.frame.
}
\description{
Extract AE results for one gene
}
