% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneASE.R
\name{getGeneASE}
\alias{getGeneASE}
\title{Extract ASE site-level results for one gene}
\usage{
getGeneASE(x, geneID, use = c("gene_name", "gene_id"))
}
\arguments{
\item{x}{An \code{ASEResult} object.}

\item{geneID}{Gene identifier or gene symbol.}

\item{use}{One of \code{"gene_name"} or \code{"gene_id"}.}
}
\value{
A data.frame.
}
\description{
Extract ASE site-level results for one gene
}
