% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneASEGene.R
\name{getGeneASEGene}
\alias{getGeneASEGene}
\title{Extract gene-level ASE results for one gene}
\usage{
getGeneASEGene(x, geneID)
}
\arguments{
\item{x}{An \code{ASEGeneResult} object.}

\item{geneID}{Gene identifier or gene symbol.}
}
\value{
A data.frame.
}
\description{
Extract gene-level ASE results for one gene
}
