% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutlierAE.R
\name{getOutlierAE}
\alias{getOutlierAE}
\title{Extract aberrant expression outliers}
\usage{
getOutlierAE(x, sampleID = NULL)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Optional character scalar specifying one sample.}
}
\value{
A data.frame containing outlier AE results.
}
\description{
Extract aberrant expression outliers
}
