% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutlierAS.R
\name{getOutlierAS}
\alias{getOutlierAS}
\title{Extract AS outliers}
\usage{
getOutlierAS(x, site = c("both", "start", "end"), sampleID = NULL)
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{site}{One of "start", "end", "both".}

\item{sampleID}{Optional sample ID.}
}
\value{
A data.frame if one site is requested, or a list if \code{site = "both"}.
}
\description{
Extract AS outliers
}
