% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutlierASE.R
\name{getOutlierASE}
\alias{getOutlierASE}
\title{Extract ASE outliers}
\usage{
getOutlierASE(x, sampleID = NULL)
}
\arguments{
\item{x}{An \code{ASEResult} object.}

\item{sampleID}{Optional sample ID.}
}
\value{
A data.frame.
}
\description{
Extract ASE outliers
}
