% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleAE.R
\name{getSampleAE}
\alias{getSampleAE}
\title{Extract AE results for one sample}
\usage{
getSampleAE(x, sampleID)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Character scalar specifying the target sample.}
}
\value{
A data.frame containing AE results for one sample.
}
\description{
Extract AE results for one sample
}
