% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleAS.R
\name{getSampleAS}
\alias{getSampleAS}
\title{Extract AS results for one sample}
\usage{
getSampleAS(x, sampleID, site = c("start", "end", "both"))
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{sampleID}{Character scalar specifying the target sample.}

\item{site}{One of "start", "end", "both".}
}
\value{
A data.frame if one site is requested, or a list if \code{site = "both"}.
}
\description{
Extract AS results for one sample
}
