% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleASE.R
\name{getSampleASE}
\alias{getSampleASE}
\title{Extract ASE results for one sample}
\usage{
getSampleASE(x, sampleID)
}
\arguments{
\item{x}{An \code{ASEResult} object.}

\item{sampleID}{Sample ID.}
}
\value{
A data.frame.
}
\description{
Extract ASE results for one sample
}
