% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleASEGene.R
\name{getSampleASEGene}
\alias{getSampleASEGene}
\title{Extract gene-level ASE results for one sample}
\usage{
getSampleASEGene(x, sampleID)
}
\arguments{
\item{x}{An \code{ASEGeneResult} object.}

\item{sampleID}{Sample ID.}
}
\value{
A data.frame.
}
\description{
Extract gene-level ASE results for one sample
}
