% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseASEReadCounter.R
\name{parseASEReadCounter}
\alias{parseASEReadCounter}
\title{Parse GATK ASEReadCounter table}
\usage{
parseASEReadCounter(x, sampleID = NULL)
}
\arguments{
\item{x}{A data.frame containing GATK ASEReadCounter results.}

\item{sampleID}{Optional sample ID. If \code{x} does not contain a sample column,
this value will be used.}
}
\value{
A standardized ASE table.
}
\description{
Parse GATK ASEReadCounter table
}
