% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFeatureCounts.R
\name{parseFeatureCounts}
\alias{parseFeatureCounts}
\title{Parse featureCounts output table}
\usage{
parseFeatureCounts(x)
}
\arguments{
\item{x}{A featureCounts result table as a data.frame.}
}
\value{
A list with counts matrix, gene length, and rowData.
}
\description{
Parse featureCounts output table
}
\keyword{internal}
