% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRegtoolsAnnotate.R
\name{parseRegtoolsAnnotate}
\alias{parseRegtoolsAnnotate}
\title{Parse regtools annotate table}
\usage{
parseRegtoolsAnnotate(x)
}
\arguments{
\item{x}{A regtools annotate result table.}
}
\value{
A standardized data.frame.
}
\description{
Parse regtools annotate table
}
\keyword{internal}
