% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AEResult.R
\name{plot.AEResult}
\alias{plot.AEResult}
\title{Plot AEResult object}
\usage{
\method{plot}{AEResult}(
  x,
  type = c("volcano", "zscore", "sample_summary", "gene", "gene_oe"),
  sampleID = NULL,
  geneID = NULL,
  groupBy = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{type}{Type of plot: "volcano", "zscore", "sample_summary", "gene", "gene_oe".}

\item{sampleID}{Required for sample-level plots; optional for gene-level plots.}

\item{geneID}{Required for gene-level plots.}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}

\item{...}{Additional arguments passed to underlying plotting functions.}
}
\value{
A ggplot object.
}
\description{
Plot AEResult object
}
