% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ASResult.R
\name{plot.ASResult}
\alias{plot.ASResult}
\title{Plot ASResult object}
\usage{
\method{plot}{ASResult}(
  x,
  type = c("zscore", "junction", "sample_summary"),
  site = c("start", "end"),
  sampleID = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  groupBy = NULL,
  ...
)
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{type}{Type of plot: "zscore", "junction", "sample_summary".}

\item{site}{One of "start" or "end".}

\item{sampleID}{Required for zscore plot; optional for junction plot.}

\item{chrom, start, end}{Required for junction plot.}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}

\item{...}{Additional arguments passed to underlying plotting functions.}
}
\value{
A ggplot object.
}
\description{
Plot ASResult object
}
