% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEGeneObservedExpected.R
\name{plotAEGeneObservedExpected}
\alias{plotAEGeneObservedExpected}
\title{Plot observed vs expected expression for one gene across samples}
\usage{
plotAEGeneObservedExpected(
  x,
  geneID,
  sampleID = NULL,
  logScale = TRUE,
  groupBy = NULL
)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{geneID}{Character scalar specifying the target gene.}

\item{sampleID}{Optional character scalar specifying one sample to highlight.}

\item{logScale}{Logical; whether to use log10 scale.}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}
}
\value{
A ggplot object.
}
\description{
Plot observed vs expected expression for one gene across samples
}
