% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAESampleSummary.R
\name{plotAESampleSummary}
\alias{plotAESampleSummary}
\title{Plot AE outlier counts per sample}
\usage{
plotAESampleSummary(x)
}
\arguments{
\item{x}{An \code{AEResult} object.}
}
\value{
A ggplot object.
}
\description{
Plot AE outlier counts per sample
}
