% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEZScore.R
\name{plotAEZScore}
\alias{plotAEZScore}
\title{Plot top AE z-scores for one sample}
\usage{
plotAEZScore(x, sampleID, topN = 30, groupBy = NULL, sample_info = NULL)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Sample ID.}

\item{topN}{Number of top genes ranked by absolute z-score to display.}

\item{groupBy}{Optional grouping column in \code{sample_info}; only used for subtitle.}

\item{sample_info}{Optional sample metadata overriding \code{x$metadata$sample_info}.}
}
\value{
A ggplot object.
}
\description{
Plot top AE z-scores for one sample
}
