% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASJunction.R
\name{plotASJunction}
\alias{plotASJunction}
\title{Plot PSI of one junction across samples}
\usage{
plotASJunction(
  x,
  chrom,
  start,
  end,
  site = c("start", "end"),
  sampleID = NULL,
  groupBy = NULL
)
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{chrom}{Chromosome.}

\item{start}{Junction start.}

\item{end}{Junction end.}

\item{site}{One of "start" or "end".}

\item{sampleID}{Optional sample to highlight.}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}
}
\value{
A ggplot object.
}
\description{
Plot PSI of one junction across samples
}
