% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASJunctionCartoon.R
\name{plotASJunctionCartoon}
\alias{plotASJunctionCartoon}
\title{Plot a cartoon junction diagram for one sample versus controls}
\usage{
plotASJunctionCartoon(
  x,
  sampleID,
  chrom,
  start,
  end,
  site = c("start", "end"),
  topN = 3
)
}
\arguments{
\item{x}{An \code{ASResult} object, preferably generated with \code{returnAll = TRUE}.}

\item{sampleID}{Character scalar specifying the target sample.}

\item{chrom}{Chromosome of the focal junction.}

\item{start}{Start coordinate of the focal junction.}

\item{end}{End coordinate of the focal junction.}

\item{site}{One of "start" or "end".}

\item{topN}{Integer; maximum number of junctions to show.}
}
\value{
A ggplot object.
}
\description{
Plot a cartoon junction diagram for one sample versus controls
}
