% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASPSI.R
\name{plotASPSI}
\alias{plotASPSI}
\title{Plot PSI for a specific splice junction}
\usage{
plotASPSI(
  x,
  chrom,
  start,
  end,
  site = c("start", "end"),
  gene = NULL,
  groupBy = NULL,
  sample_info = NULL,
  decreasing = TRUE
)
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{chrom}{Chromosome.}

\item{start}{Junction start position.}

\item{end}{Junction end position.}

\item{site}{One of \code{"start"} or \code{"end"}.
Determines whether to use \code{result_start} or \code{result_end}.}

\item{gene}{Optional gene name to further filter events.}

\item{groupBy}{Optional grouping column in \code{sample_info}.}

\item{sample_info}{Optional data.frame; overrides \code{x$metadata$sample_info}.}

\item{decreasing}{Logical; whether to sort samples by PSI.}
}
\value{
A ggplot object.
}
\description{
Visualize PSI values across samples for a given splice junction,
defined by chrom + start + end.
}
