% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASZscore.R
\name{plotASZScore}
\alias{plotASZScore}
\title{Plot AS z-score ranking for one sample}
\usage{
plotASZScore(x, sampleID, site = c("start", "end"), topN = 10)
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{sampleID}{Character scalar specifying the target sample.}

\item{site}{One of "start" or "end".}

\item{topN}{Integer; number of top junctions to label.}
}
\value{
A ggplot object.
}
\description{
Plot AS z-score ranking for one sample
}
