% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExonAnnotation.R
\name{readExonAnnotation}
\alias{readExonAnnotation}
\title{Read exon annotation table}
\usage{
readExonAnnotation(file)
}
\arguments{
\item{file}{Path to a tab-delimited exon annotation file.}
}
\value{
A standardized exon annotation data.frame.
}
\description{
Read exon annotation table
}
