% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAS.R
\name{runAS}
\alias{runAS}
\title{Run aberrant splicing detection}
\usage{
runAS(
  x,
  mode = c("both", "start", "end"),
  pAdjustMethod = "bonferroni",
  pCutoff = 0.05,
  minTotalReads = 10,
  minJunctionReads = 5,
  deltaPSI_cutoff = 0.1,
  sample_info = NULL,
  returnAll = FALSE
)
}
\arguments{
\item{x}{A regtools annotate table.}

\item{mode}{One of "both", "start", "end".}

\item{pAdjustMethod}{Multiple testing correction method.}

\item{pCutoff}{Adjusted p-value cutoff.}

\item{minTotalReads}{Minimum total reads at start/end site for defining outliers.}

\item{minJunctionReads}{Minimum reads supporting focal junction for defining outliers.}

\item{deltaPSI_cutoff}{Numeric; threshold for defining aberrant splicing events.
Default is 0.1 (10\%). Events with absolute delta PSI ≥ this cutoff are considered aberrant.}

\item{sample_info}{Optional sample metadata as a data.frame.
Must contain a column \code{sample_id}, and can include additional annotations
such as \code{sex}, \code{Gene_diagnostic}, or user-defined grouping variables.}

\item{returnAll}{Logical; whether to return all tested junctions. If FALSE,
only outliers are returned.}
}
\value{
An object of class \code{ASResult}.
}
\description{
Perform junction-centric aberrant splicing analysis using regtools annotate results.
}
