% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runASE.R
\name{runASE}
\alias{runASE}
\title{Run site-level ASE analysis}
\usage{
runASE(
  x,
  annotation = NULL,
  gtf = NULL,
  minDP = 10,
  region_keep = c("exonic", "UTR"),
  pCutoff = 0.05,
  ratioCutoff = 0.7,
  sex = NULL,
  chrX = c("X", "chrX"),
  returnAll = FALSE
)
}
\arguments{
\item{x}{A merged ASEReadCounter table.}

\item{annotation}{Optional transcript annotation list with components \code{exon} and \code{utr}.}

\item{gtf}{Optional local GTF path. Used only if \code{annotation} is not provided.}

\item{minDP}{Minimum total read depth.}

\item{region_keep}{Regions to retain, default \code{c("exonic", "UTR")}.}

\item{pCutoff}{Adjusted p-value cutoff.}

\item{ratioCutoff}{Allelic ratio cutoff. Sites with \code{alt_ratio > ratioCutoff}
or \verb{< 1 - ratioCutoff} are considered significant.}

\item{sex}{Optional sex information. Can be either:
\itemize{
\item a named character vector, e.g. \code{c(S1 = "female", S2 = "male")}
\item a data.frame with columns \code{sample_id} and \code{sex}
}
If not provided, chrX sites are excluded by default.}

\item{chrX}{Character vector specifying chromosome X names, default \code{c("X", "chrX")}.}

\item{returnAll}{Logical; whether to return all tested sites. If FALSE, only outliers are returned.}
}
\value{
An object of class \code{ASEResult}.
}
\description{
Run site-level ASE analysis
}
