% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-AEResult.R
\name{summary.AEResult}
\alias{summary.AEResult}
\title{Summarize AEResult object}
\usage{
\method{summary}{AEResult}(object, ...)
}
\arguments{
\item{object}{An \code{AEResult} object.}

\item{...}{Reserved.}
}
\value{
A summary object.
}
\description{
Summarize AEResult object
}
