% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ASEGeneResult.R
\name{summary.ASEGeneResult}
\alias{summary.ASEGeneResult}
\title{Summary of gene-level ASE results}
\usage{
\method{summary}{ASEGeneResult}(object, bySample = TRUE, ...)
}
\arguments{
\item{object}{An \code{ASEGeneResult} object.}

\item{bySample}{Logical; whether to compute per-sample summary.}
}
\value{
A list containing summary statistics.
}
\description{
Summary of gene-level ASE results
}
