% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-classed-ASE.R
\name{summary.ASEResult}
\alias{summary.ASEResult}
\title{Summary of ASE results}
\usage{
\method{summary}{ASEResult}(object, bySample = TRUE, ...)
}
\arguments{
\item{object}{An \code{ASEResult} object.}

\item{bySample}{Logical; whether to compute per-sample summary.}
}
\value{
A list containing summary statistics.
}
\description{
Summary of ASE results
}
