% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ASResult.R
\name{summary.ASResult}
\alias{summary.ASResult}
\title{Summarize ASResult object}
\usage{
\method{summary}{ASResult}(object, ...)
}
\arguments{
\item{object}{An \code{ASResult} object.}

\item{...}{Reserved.}
}
\value{
A summary object of class \code{summary.ASResult}.
}
\description{
Summarize ASResult object
}
