% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyAS.R
\name{tidyAS}
\alias{tidyAS}
\title{Return tidy AS result table}
\usage{
tidyAS(x, site = c("start", "end", "both"))
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{site}{One of "start", "end", "both".}
}
\value{
A data.frame if one site is requested, or a list if \code{site = "both"}.
}
\description{
Return tidy AS result table
}
