% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeExonAnnotation.R
\name{writeExonAnnotation}
\alias{writeExonAnnotation}
\title{Write exon annotation table}
\usage{
writeExonAnnotation(x, file)
}
\arguments{
\item{x}{Exon annotation data.frame.}

\item{file}{Output file path.}
}
\value{
Invisibly returns \code{file}.
}
\description{
Write exon annotation table
}
