% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneAS.R
\name{getGeneAS}
\alias{getGeneAS}
\title{Extract AS results for one gene}
\usage{
getGeneAS(x, geneID, site = c("both", "start", "end"), use = "gene_names")
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{geneID}{Gene identifier or gene symbol.}

\item{site}{One of \code{"start"}, \code{"end"}, or \code{"both"}.}

\item{use}{Column used for matching, usually \code{"gene_names"} or \code{"gene_ids"}.}
}
\value{
A data.frame if one site is requested, or a list if \code{site = "both"}.
}
\description{
Extract AS results for one gene
}
