% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEGene.R
\name{plotAEGene}
\alias{plotAEGene}
\title{Plot expression pattern of one gene across samples}
\usage{
plotAEGene(
  x,
  geneID,
  sampleID = NULL,
  value = c("raw_count", "norm_count", "expected_count", "zscore", "log2fc"),
  logScale = TRUE,
  groupBy = NULL
)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{geneID}{Character scalar specifying the target gene.}

\item{sampleID}{Optional character scalar specifying one sample to highlight.}

\item{value}{Which value to plot. One of "raw_count", "norm_count", "expected_count", "zscore", "log2fc".}

\item{logScale}{Logical; whether to use log10 scale for count-based values.}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}
}
\value{
A ggplot object.
}
\description{
Plot expression pattern of one gene across samples
}
