% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEGene.R
\name{plotAEGene}
\alias{plotAEGene}
\title{Plot AE values for one gene across samples}
\usage{
plotAEGene(
  x,
  geneID,
  use = c("gene_id", "gene_name"),
  value = c("raw_count", "norm_count", "expected_count", "zscore", "log2fc"),
  logScale = FALSE,
  groupBy = NULL,
  sample_info = NULL
)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{geneID}{Gene identifier or gene symbol.}

\item{use}{One of \code{"gene_id"} or \code{"gene_name"}.}

\item{value}{Which value to plot. One of
\code{"raw_count"}, \code{"norm_count"}, \code{"expected_count"}, \code{"zscore"}, \code{"log2fc"}.}

\item{logScale}{Logical; whether to use log10 scale for count-like values.}

\item{groupBy}{Optional grouping column in \code{sample_info}.}

\item{sample_info}{Optional sample metadata overriding \code{x$metadata$sample_info}.}
}
\value{
A ggplot object.
}
\description{
Plot AE values for one gene across samples
}
