% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEGeneObservedExpected.R
\name{plotAEGeneObservedExpected}
\alias{plotAEGeneObservedExpected}
\title{Plot observed vs expected expression for one gene}
\usage{
plotAEGeneObservedExpected(
  x,
  geneID,
  use = c("gene_id", "gene_name"),
  value = c("norm_count", "raw_count"),
  logScale = FALSE,
  groupBy = NULL,
  sample_info = NULL
)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{geneID}{Gene identifier or symbol.}

\item{use}{\code{"gene_id"} or \code{"gene_name"}.}

\item{value}{\code{"norm_count"} or \code{"raw_count"}}

\item{logScale}{Whether to log-scale y axis.}

\item{groupBy}{Grouping column.}

\item{sample_info}{Optional override metadata.}
}
\value{
ggplot object
}
\description{
Plot observed vs expected expression for one gene
}
