% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAESampleSummary.R
\name{plotAESampleSummary}
\alias{plotAESampleSummary}
\title{Plot AE outlier burden per sample}
\usage{
plotAESampleSummary(x, groupBy = NULL, sample_info = NULL, decreasing = TRUE)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{groupBy}{Optional grouping column in \code{sample_info}.}

\item{sample_info}{Optional sample metadata overriding \code{x$metadata$sample_info}.}

\item{decreasing}{Logical; whether to sort in decreasing order of outlier burden.}
}
\value{
A ggplot object.
}
\description{
Plot AE outlier burden per sample
}
