% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEVolcano.R
\name{plotAEVolcano}
\alias{plotAEVolcano}
\title{Plot AE volcano plot for one sample}
\usage{
plotAEVolcano(x, sampleID, padjCutoff = 0.05, zCutoff = NULL, labelTopN = 10)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Sample ID.}

\item{padjCutoff}{FDR cutoff for highlighting outliers.}

\item{zCutoff}{Optional z-score cutoff.}

\item{labelTopN}{Number of top genes (by absolute z-score) to label.}
}
\value{
A ggplot object.
}
\description{
Plot AE volcano plot for one sample
}
