% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEVolcano.R
\name{plotAEVolcano}
\alias{plotAEVolcano}
\title{Plot AE volcano plot for one sample}
\usage{
plotAEVolcano(
  x,
  sampleID,
  padjCutoff = 0.05,
  zCutoff = NULL,
  labelTopN = 10,
  groupBy = NULL,
  sample_info = NULL
)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Sample ID.}

\item{padjCutoff}{FDR cutoff for highlighting outliers.}

\item{zCutoff}{Optional z-score cutoff.}

\item{labelTopN}{Number of top genes (by |zscore|) to label.}

\item{groupBy}{Optional grouping column (only used in subtitle).}

\item{sample_info}{Optional sample metadata overriding stored metadata.}
}
\value{
ggplot object
}
\description{
Plot AE volcano plot for one sample
}
