% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEVolcano.R
\name{plotAEVolcano}
\alias{plotAEVolcano}
\title{Plot AE volcano plot for one sample}
\usage{
plotAEVolcano(x, sampleID, topN = 10)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Character scalar specifying the target sample.}

\item{topN}{Integer; number of top genes to label.}
}
\value{
A ggplot object.
}
\description{
Plot AE volcano plot for one sample
}
