% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEZScore.R
\name{plotAEZScore}
\alias{plotAEZScore}
\title{Plot AE z-score ranking for one sample}
\usage{
plotAEZScore(x, sampleID, topN = 10, groupBy = NULL)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Character scalar specifying the target sample.}

\item{topN}{Integer; number of top genes to label.}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}
}
\value{
A ggplot object.
}
\description{
Plot AE z-score ranking for one sample
}
