% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAEZScore.R
\name{plotAEZScore}
\alias{plotAEZScore}
\title{Plot top AE z-scores for one sample}
\usage{
plotAEZScore(x, sampleID, topN = 30)
}
\arguments{
\item{x}{An \code{AEResult} object.}

\item{sampleID}{Sample ID.}

\item{topN}{Number of top genes ranked by absolute z-score to display.}
}
\value{
A ggplot object.
}
\description{
Plot top AE z-scores for one sample
}
