% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotASSampleSummary.R
\name{plotASSampleSummary}
\alias{plotASSampleSummary}
\title{Plot AS outlier counts per sample}
\usage{
plotASSampleSummary(x, site = c("start", "end"), groupBy = NULL)
}
\arguments{
\item{x}{An \code{ASResult} object.}

\item{site}{One of "start" or "end".}

\item{groupBy}{Optional character scalar specifying a column name in stored \code{colData}.}
}
\value{
A ggplot object.
}
\description{
Plot AS outlier counts per sample
}
