% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAS.R
\name{runAS}
\alias{runAS}
\title{Run aberrant splicing detection}
\usage{
runAS(
  x,
  mode = c("both", "donor", "acceptor"),
  pAdjustMethod = "bonferroni",
  pCutoff = 0.05,
  minTotalReads = 10,
  minJunctionReads = 5
)
}
\arguments{
\item{x}{A regtools annotate table.}

\item{mode}{One of "both", "donor", "acceptor".}

\item{pAdjustMethod}{Multiple testing correction method.}

\item{pCutoff}{Adjusted p-value cutoff.}

\item{minTotalReads}{Minimum total reads at donor/acceptor site.}

\item{minJunctionReads}{Minimum reads supporting focal junction.}
}
\value{
An object of class \code{ASResult}.
}
\description{
Perform junction-centric aberrant splicing analysis using regtools annotate results.
}
